RA-26 clean up of unused code and filenames
