+++ src/SFA.Apprenticeships.Application.Vacancy/IVacancyManagementService.cs
@@ -8,19 +8,4 @@
         IServiceResult Delete(int vacancyId);
         IServiceResult<VacancySummary> FindSummary(int vacancyId);
     }
-
-    public class VacancyManagementServiceCodes
-    {
-        public class Delete
-        {
-            public const string Ok = "VacancyManagement.Delete.Ok";
-            public const string VacancyInIncorrectState = "Vacancy.Delete.VacancyInIncorrectState";
-        }
-
-        public class FindSummary
-        {
-            public const string Ok = "VacancyManagement.FindSummary.Ok";
-            public const string NotFound = "VacancyManagement.FindSummary.NotFound";
-        }
-    }
 }
