+++ src/SFA.Apprenticeships.Application.Vacancy/StrategyResult{T}.cs
